+++ src/Action/RouteCapabilityOverview.php
@@ -5,10 +5,10 @@ namespace Heptacom\HeptaConnect\Storage\ShopwareDal\Action;
 
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\FetchMode;
-use Doctrine\DBAL\Query\QueryBuilder;
 use Heptacom\HeptaConnect\Storage\Base\Contract\RouteCapabilityOverviewActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\RouteCapabilityOverviewCriteria;
 use Heptacom\HeptaConnect\Storage\Base\Contract\RouteCapabilityOverviewResult;
+use Heptacom\HeptaConnect\Storage\ShopwareDal\Support\Query\QueryBuilder;
 use Shopware\Core\Defaults;
 
 class RouteCapabilityOverview implements RouteCapabilityOverviewActionInterface
@@ -83,7 +83,7 @@ class RouteCapabilityOverview implements RouteCapabilityOverviewActionInterface
 
     protected function getBuilder(): QueryBuilder
     {
-        $builder = $this->connection->createQueryBuilder();
+        $builder = new QueryBuilder($this->connection);
 
         // TODO human readable
         return $builder