+++ test/Action/RouteFindByTargetsAndTypeTest.php
@@ -15,6 +15,7 @@ use Shopware\Core\Framework\Uuid\Uuid;
 
 /**
  * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\Action\RouteFindByTargetsAndType
+ * @covers \Heptacom\HeptaConnect\Storage\ShopwareDal\Support\Query\QueryBuilder
  */
 class RouteFindByTargetsAndTypeTest extends TestCase
 {